add changes to RDB

+++ Models/PODetails.cs
@@ -16,7 +16,7 @@ namespace jvPo.Models
     public int Quantity { get; set; }
     public string Unit { get; set; } = string.Empty;
     public string Description { get; set; } =string.Empty;
-    public double Price { get; set; }
-    public double Total { get; set; }  
+    public decimal Price { get; set; }
+    public decimal Total { get; set; }  
     }
 }
